MDW-3: Download first chapter for bleach
Add code to support crawling through the first page
of a chapter in order to get the number of pages
in the chapter.
Add code to crawl to the image location for the page.
Add code to download the image and save it in a new
file on the disk.